+++ aliases.js
@@ -1,0 +1,9 @@
+const path = require('path');
+const tsconfig = require('./tsconfig');
+
+const { compilerOptions: { baseUrl, paths } } = tsconfig;
+const aliasReg = (str) => str.replace(/^(.*)\/\*$/, '$1');
+module.exports = Object.keys(paths).reduce(
+    (obj, a) => (obj[aliasReg(a)] = path.resolve(__dirname, aliasReg(`${baseUrl}/${paths[a]}`)), obj),
+    {}
+);