+++ src/@classes/public/IsometricCanvas/types.ts
@@ -1,4 +1,5 @@
 export interface IsometricCanvasProps {
+    container?: HTMLElement | string;
     backgroundColor?: string;
     scale?: number;
     height?: number;